requirements.txt
apscheduler # Sceduler
tk # GUI
pyserial # Serial communication
pillow # Image processing